Inventerings sida med fixat fel och lista över hela skärmen.
Vi har gjort en ingrediens klass och en inventerings aktivitet med en adapter. Just nu är inventerings sidan bunden till button 3 på startsidan. FIxade crashen från förra commiten. Just nu går listan över hela skärmen

+++ app/src/main/java/se/miun/antonsskafferi/Ingredient.java
@@ -6,10 +6,10 @@ package se.miun.antonsskafferi;
 
 public class Ingredient {
     private String name;
-    private int quantity;
+    private double quantity;
     private String unit;
 
-    public Ingredient(String name, int quantity, String unit){
+    public Ingredient(String name, double quantity, String unit){
         this.name = name;
         this.quantity = quantity;
         this.unit = unit;
@@ -19,7 +19,7 @@ public class Ingredient {
         return name;
     }
 
-    public int getQuantity(){
+    public double getQuantity(){
         return quantity;
     }
